only print secret keys in build summary output
